Debugging

+++ AgroVia/src/main/java/com/agrosolutions/AgroVia/dto/ResourceRequest.java
@@ -9,8 +9,8 @@ public class ResourceRequest {
     private BigDecimal pricePerWeek;
     private BigDecimal depositAmount;
     private String location;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private Integer deliveryRadius;
     private String brand;
     private String model;
@@ -90,19 +90,19 @@ public class ResourceRequest {
         this.location = location;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
